Gave the ChemistrySolver() a method to solve the concentrations for a certain timestep.
You pass solveConcentrationsAfterTime() the starting concentrations and the
length of the timestep, and it returns the concentrations at the end of that
timestep. Also gave an example of how to use it in the if __name__=="__main__"
section. Notice it's slower than solving the whole thing in one go, as you have
to keep restarting the ode solver - but there's probably not a lot we can do
about that.