setConcentrations() now accepts array or dictionary
This is so you can set concentrations of species you know about (eg the fuel)
without having to know what all the other species are.

The internally stored self.concentrations that are set this way are not currently used though - but you could access them with a getConcentrations() if you wanted.