Use Mintz's correlation for enthalpy of solvation in linear alkanes
This H is now used to calculate the S from the G, rather than before
when Pierotti's S was used to calculate the H from the G.

The old methods are still there and can be called explicitly if required.

see equation 15 in 
Enthalpy of Solvation Correlations For Gaseous Solutes 
Dissolved in Linear Alkanes (C5 - C16) Based on the Abraham Model
C Mintz, K Burton, WE Acree Jr, MH Abraham
http://dx.doi.org/10.1002/qsar.200730040


This gives a much smaller dependence of partition coefficient on temperature,
and as a result the evaporation simulation behaves better.