Created class for DepositPhase, with an equilibrate method, in evapor.py
The equilibrate method may be wrong - we might have to do this with 
concentrations rather than amounts. I guess I'm hoping that we can just 
assume that everything has the same total molar density and it will 
all cancel out, or something like that. I haven't thought it through 
as thoroughly as it deserves though.