Added Amrit's diesel RMG results to RMG_results-amrit
The currently active results have 73 species in the core. I don't trust the
150-species Chemkin file much because something strange happened for there 
to be no 74 through 149 files!

This was generated with both Pfaendtner and Broadbelts rules, and RMG's rules.
There are numerous undeclared duplicate reactions.

NOTICE that the fuel components have different names and are in a different order
than in the mechanisms we've been working with recently.