Chemistry-only simulation, runs for 1 nanosecond.
This shows how fast *some* of the reactions are. Advancing more than
a picosecond at the start causes DASSL to complain it's taking too 
many steps. The temporary solution is to run the simulation for 100 steps 
across 1 nanosecond, but on that time-scale not much else happens.

Better would be to change how the ODE solver takes steps, so it can start
off stepping as small as it likes.

Oh, and the oxygen concentration is (still) going negative.