DASSL takes as many steps as it needs, not just 500.
I updated the PyDAS project, so remember to update your submodule 
and re-build your Cython (python setup.py build_ext --inplace)

In theory, the simulation can now do a millisecond or more,
not just a picosecond, but in practice I have got bored 
waiting for it at about 155 picoseconds. 

This is a problem!