Changed how oxygen dissolution is modeled.
I think directly changing the amount vector was screwing up
the DAE solver. Now there's an additional term in the ODEs
for oxygen dissolving into the liquid. This means we have to
estimate another characteristic time scale, but this can 
be found from diffusion or something (for now I have k=1e2 s^-1)